[change] Devide packages: client -> {client, install}
New pkg client is HTTP client.
New pkg install is for installation.
