+++ install/fetch.go
@@ -1,4 +1,4 @@
-package client
+package install
 
 import (
 	"encoding/hex"
@@ -10,6 +10,7 @@ import (
 	"path"
 	"path/filepath"
 
+	"github.com/binqry/binq/client"
 	"github.com/binqry/binq/internal/erron"
 	"github.com/binqry/binq/schema/item"
 )
@@ -18,7 +19,7 @@ func (r *Runner) fetch() (err error) {
 	if r.sourceURL == "" {
 		return fmt.Errorf("Can't fetch because sourceURL is not set. Source: %s", r.Source)
 	}
-	req, err := NewHttpGetRequest(r.sourceURL, map[string]string{})
+	req, err := client.NewHttpGetRequest(r.sourceURL, map[string]string{})
 	if err != nil {
 		return err
 	}
