+++ install/install.go
@@ -1,0 +1,2 @@
+// Package install implements installation functionality of binq.
+package install
